dtype fixup

+++ pandas/tests/groupby/test_nunique.py
@@ -191,7 +191,7 @@ def test_empty_categorical(observed):
     gb = ser.groupby(ser, observed=observed)
     result = gb.nunique()
     if observed:
-        expected = Series([], index=cat[:0], dtype=int)
+        expected = Series([], index=cat[:0], dtype="int64")
     else:
-        expected = Series([0], index=cat)
+        expected = Series([0], index=cat, dtype="int64")
     tm.assert_series_equal(result, expected)
